Copy embed tag instead of asking for embed code
In this branch, we're copying the HTML for the embed right from the live
page and sticking it in the popout window. This is cool because it lets
us keep the extra controls like quality and CC.

This mostly works great. There is probably some lint to be cleaned
(messing with the #movie_player object more than we need to).

For some reason, when this thing loads it clobbers the rest of the page.
It won't respect my margin and padding settings, and when I inspect the
document, I find that it has rewritten the whole DOM. Or something. Most
of my page elements are missing, including the div which is supposed to
wrap around the player and script tags. Chrome reports that the
margin-top:8px is coming from a client stylesheet, which is nonsense. I
suspect that the JavaScript that is coming from $('#watch-player').html()
is rewriting the page. Or something.

I interrupt the process by sticking an alert or something in there,