Some changes to fit new Chrome store.
Also, the Chrome Developer thing is acting crazy with locales. It's
displaying Spanish when English is selected and vice versa. By
repeatedly uploading the same package with an incremented version
number, I got it to display English for English, but now it also
displays English for Spanish. I'll try pushing a 3.4.5 or so in the next
couple days and see if that fixes it. Goddammit.

https://chrome.google.com/webstore/detail/pofekaindcmmojfnfgbpklepkjfilcep?hl=en
https://chrome.google.com/webstore/detail/pofekaindcmmojfnfgbpklepkjfilcep?hl=es