Window is opened in panel instead of popup
Video window is opened as Chrome panel - attached to lower right part
of screen and on top of other windows. Panel can be detached from
position and moved around the screen, acting as Chrome popup window
(not being always on top). Panel window can be attached back to initial
position (always on top).